pyproject.toml
[tool.poetry]
name = "django-proj"
version = "0.1.0"
description = ""
authors = ["Cavinton <enviuslxrd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
